[PATCH] ppc64: Fix thinko in prom_init.c
(It's a real bug, but I suspect it doesn't trigger normally as we tend
to allocate the initrd low, but it should be fixed anyway).

This patch fixes an error in prom_init.c in the check for the initrd
location vs. the memory allocation mecanism.

Signed-off-by: Arnd Bergmann <arndb@de.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4229edd0kw4tMSjIckhv-kH4LnqMCw