Merge pobox.com:/garz/repo/netdev-2.6/wan-fixes into pobox.com:/garz/repo/net-drivers-2.6
2005/03/03 00:22:23-05:00 akpm
[PATCH] drivers/net/wan/z85230.c interrupt handling fix

From: tom watson <lkml@tommywatson.com>

While working on a driver for z85230 based board I noticed what looks like
it could be a problem.  If the interrupt handler is handling an interrupt
on port b and an interrupt comes in for port a, it seems to me that the
port b handler would be called instead of the port a handler, and possibly
hang the board until reset.

Attached is a patch to set the irq methods back to port a before attempting
to call them.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

2005/03/03 00:19:59-05:00 khc
[PATCH] WAN drivers fix: N2, C101, PCI200SYN - quite fatal

The last change to drivers/wan/hd6457x.c was a bit fatal to drivers
using it - the attached patch fixes NULL pointer dereference on RX.

I've updated URLs in MAINTAINERS and wan/Kconfig as well.

Signed-off-by: Krzysztof Halasa <khc@pm.waw.pl>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

BKrev: 422a3c00MVq61CZ1TUIIh1Ot50WZFA