Merge bk://kernel.bkbits.net/davem/net-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2005/03/07 09:34:59-08:00 herbert
[IPSEC]: Kill redundant dst_release check in xfrm_dst_destroy

Here's a trivial patch to get rid of a redundant check that I added
in patch 3/4.  dst_release already checks for dst == NULL.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/07 09:33:29-08:00 akpm
[ATALK]: linux/atalk.h needs net/sock.h

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/07 09:32:25-08:00 herbert
[IPSEC]: Get rid of dst_pmtu/ext2_header_len

Here is a patch that replaces all occurrences of dst_pmtu in the TCP
stack.  As a result we no longer need ext2_header_len.

This has a nice synergetic effect with Arnaldo's latest change to
linux/tcp.h :)

I'll be removing other users of dst->path/dst_pmtu next.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/07 09:31:04-08:00 herbert
[IPSEC]: Check dst validity harder in xfrm_bundle_ok

There is another bug in xfrm_bundle_ok where I forgot to
check the validity of xdst->route.  In fact, the check
on dst->path isn't strong enough either.  For IPv6 entries,
dst->path->obsolete is always negative until you call
ipv6_dst_check.  So we really need to do that here.

Here's the patch to fix those two problems.  Yes I know
my dst_check implementation is lame.  I'll come back and
fix up all the dst_check functions by moving their dst_release
calls out.  It proves that you were right in that IPv6 dst
leak thread :)

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/06 21:36:51-08:00 herbert
[IPSEC]: Get metrics for xfrm_dst from top dst

For the metrics other than MTU, we should also be getting it from
the top dst as opposed to the bottom dst.  With this change, the
user is able to manipulate values such as advmss for individual
hosts behind a remote IPsec gateway.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/06 21:35:27-08:00 herbert
[IPSEC]: Fix xfrm[46]_update_pmtu to update top dst

Let's also fix IPsec PMTU storage.  When we get an MTU update for an
xfrm_dst, it should be done to the top dst, not the bottom dst.

For example, when we get a need-to-frag message for host C behind
a our IPsec peer B, we should be updating the dst entry for C and
not B as we do now.

I've removed the boundary checks since the same checks are done
in ipv[46]/route.c already.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/06 21:34:23-08:00 herbert
[IPSEC]: Use dst_mtu i n xfrm[46]_output

There was a lot of fun in laying the foundations.  Now it's time
to start using it.

This patch fixes the calculations in xfrm[46]_output so that we don't
reject properly sized packets from the TCP stack.  The previous
calculation is wrong because the xfrm overhead is not constant.

After this I'll start cleaning stuff up.  For example, tcp's
ext2_header_len and dst's path can both be removed.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/06 21:31:52-08:00 herbert
[IPSEC]: Store MTU at each xfrm_dst

Finally this is the patch that sets the MTU values of each xfrm_dst
and keeps it up-to-date.

To recap, at this point we've obtained accurate MTU values at each
xfrm_dst.  The next step would be to start using it as
dst_pmtu(xfrm_dst).

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/06 21:27:44-08:00 herbert
[IPSEC]: Add route element to xfrm_dst

This patch adds a pointer to the route corresponding to the specific
flow over the SA of an xfrm_dst that's being used.

It also sets the next pointer of each xfrm_dst to the one above it.
This allows to traverse the list upwards from the bottom.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/06 21:24:19-08:00 herbert
[IPSEC]: Add xfrm_state_mtu()

This patch introduces the function xfrm_state_mtu which calculates
the MTU under a specific SA.  This function can be simplified once
we get rid all other uses of get_max_size as we can move the calculation
into the invidual SA type.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/06 21:03:21-08:00 herbert
[CRYPTO]: Mark myself as co-maintainer.

James is stepping down for a while.

Signed-off-by: David S. Miller <davem@davemloft.net>

BKrev: 422c99c3V-9N9jSpp1j5u4Wp9ALG_A