[PATCH] x86: refactor memory setup
Refactor the i386 default and CONFIG_DISCONTIG_MEM setup_memory() functions to
share the common bootmem initialisation code.  This code is intended to be
identical, but there are currently some fixes applied to one and not the
other.  This patch extracts this common initialisation code.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Signed-off-by: Dave Hansen <haveblue@us.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 422de2f7blZpAnc8S19aLOdIB_XMrg