[PATCH] Fix race between the NMI code and the CMOS clock
This patch fixes a race between the CMOS clock setting and the NMI code.
The NMI code indiscriminatly sets index registers and values in the same
place the CMOS clock is set.  If you are setting the CMOS clock and an NMI
occurs, Bad values could be written to or read from the CMOS RAM, or the
NMI operation might not occur correctly.

Fixing this requires creating a special lock so the NMI code can know its
CPU owns the lock an "do the right thing" in that case.

This was discovered and the fix has been tested by a very demanding
customer who tests the heck of out the software we deliver.

Signed-off-by: Corey Minyard <minyard@acm.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 422de54edYxNk342x8lswWb4G8juHw