[PATCH] rework core barrier support
This reworks the core barrier support to be a lot nicer, so that all the
nasty code resides outside of drivers/ide.  It requires minimal changes to
support in a driver, I've added SCSI support as an example.  The ide code
is adapted to the new code.

With this patch, we support full barriers on sata now.  Bart has acked the
addition to -mm, I would like for this to be submitted as soon as 2.6.12
opens.

Signed-off-by: Jens Axboe <axboe@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 422de59d77wAY_ZNjA7k34YSFKZF1w