[PATCH] efi: fix failure handling
The EFI driver allocates memory and writes into it without checking the
success of the allocation.  Furthermore, on failure of the
firmware_register() it doesn't free the allocated memory and on failure of
the subsys_create_file() calls it returns zero instead of the errorcode.

Signed-off-by: Panagiotis Issaris <panagiotis.issaris@mech.kuleuven.ac.be>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 422de829ewVq3O_R55yGAK4tr76YTw