[PATCH] posix-timers: tidy up clock interfaces and consolidate dispatch logic
This patch cleans up the posix-timers interfaces for defining clocks, and the
calls to them.  It fixes some sloppy types, adds a clockid_t parameter to the
calls that lacked it, and adds a function pointer that can be used for
clock_getres.  It further cleans up the posix-timers.c code using the k_clock
function pointers or default functions when no hooks are supplied,
consolidating repeated code into shared inline functions or macros.  This
paves the way for adding the CPU clock hooks.

The mmtimer.c changes are untested, but obviously can't be wrong.  There
aren't any other struct k_clock definitions in the tree, but any others would
need to be updated for the function signature changes.

Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 422debccuwU4XpjWg_VarVJ5oOtbaQ