[PATCH] PANIC in check_process_timers()
It was intended that such things would not be possible because getting into
that code in the first place should be ruled out while exiting.  That
removes the requirement for any special case check in the common path.
But, it was done too late since it hadn't occurred to me that ->live going
zero itself created a problem.

Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 422dec19ZDpwogbFERiziEu6BP71OA