[PATCH] ppc64: error code cleanups for rtas wrappers
This patch is from John Rose <johnrose@austin.ibm.com>

This patch changes the rtas wrapper functions in rtas.c to map RTAS
failure codes to conventional error values.  The goal is to make
failure conditions obvious in the wrapper functions and in the caller
code.

Signed-off-by: John Rose <johnrose@austin.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 422dfc761x2YRJRFRvwnhbXeZrjqjw