[PATCH] x86: do not unnecessarily memset the pgdats
Both the pgdats and the struct zonelist are zeroed unnecessarily.  The
zonelist is a member of the pgdat, so any time the pgdat is cleared, so is the
zonelist.  All of the architectures present a zeroed pgdat to the generic
code, so it's not necessary to set it again.

Not clearing it like this allows the functions to be reused by the memory
hotplug code.  The only architecture which has a dependence on these clears is
i386.  The previous patch in this series fixed that up.

Signed-off-by: Dave Hansen <haveblue@us.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4230761758K5xP7AfJjt65A5fOmLzQ