[PATCH] cciss: per disk queue support
This patch adds per disk queue functionality.  It seems that the 2.6 kernel
expects a queue per disk.  If you have multiple logical drives on a controller
all of the queues actually point back to the same queue.  If a drive is
deleted it blows us out of the water.

We hold the lock during any queue operations and have added what we call a
"fair-enough" algorithm to prevent starving out any drive.

Signed-off-by: Mike Miller <mike.miller@hp.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4230794aPScyTRGtj5Mxu0koeYE97w