[PATCH] nfsd4: remove st_vfs_file checks
The st_vfs_file field *must* be non-NULL as long as the stateid is in use, so
there's no point testing for it, unless we want to just BUG().  But since we'd
immediately hit a NULL dereference in every case, there's no need even for
that.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Neil Brown <neilb@cse.unsw.edu.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 42307bafA2shIf92qMZiO3rZDWYgbQ